+++ fixtures/architect/src/http/any-catchall/app/routes/500.js
@@ -1,0 +1,13 @@
+export function meta() {
+  return { title: "Shoot..." };
+}
+
+export default function FiveHundred() {
+  console.error("Check your server terminal output");
+
+  return (
+    <div>
+      <h1>500</h1>
+    </div>
+  );
+}