+++ fixtures/architect/src/http/any-catchall/data/global.js
@@ -1,0 +1,5 @@
+exports.loader = async () => {
+  return {
+    date: new Date()
+  };
+};