+++ fixtures/architect/src/http/any-catchall/remix.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  appDirectory: "app",
+  dataDirectory: "data",
+  browserBuildDirectory: "../../../public/remix",
+  publicPath: "/remix/",
+  serverBuildDirectory: "build",
+  devServerPort: 8002
+};