+++ fixtures/vercel/remix.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  appDirectory: "app",
+  dataDirectory: "data",
+  browserBuildDirectory: "./public/build",
+  publicPath: "/build/",
+  serverBuildDirectory: "build",
+  devServerPort: 8002
+};