+++ jest.config.js
@@ -1,6 +1,11 @@
 module.exports = {
   projects: [
     {
+      displayName: "architect",
+      testEnvironment: "node",
+      testMatch: ["<rootDir>/packages/architect/**/*-test.[jt]s?(x)"]
+    },
+    {
       displayName: "cli",
       testEnvironment: "node",
       testMatch: ["<rootDir>/packages/cli/**/*-test.[jt]s?(x)"]
@@ -9,6 +14,16 @@ module.exports = {
       displayName: "core",
       testEnvironment: "node",
       testMatch: ["<rootDir>/packages/core/**/*-test.[jt]s?(x)"]
+    },
+    {
+      displayName: "express",
+      testEnvironment: "node",
+      testMatch: ["<rootDir>/packages/express/**/*-test.[jt]s?(x)"]
+    },
+    {
+      displayName: "vercel",
+      testEnvironment: "node",
+      testMatch: ["<rootDir>/packages/vercel/**/*-test.[jt]s?(x)"]
     },
     {
       displayName: "gists-app",