+++ packages/core/__tests__/build-test.ts
@@ -127,19 +127,19 @@ describe.skip("building", () => {
       expect(getFilenames(output)).toMatchInlineSnapshot(`
         Array [
           "_shared/Shared-b97a0af4.js",
-          "entry-server-4bebe6c1.js",
-          "pages/one-53b19748.js",
-          "pages/two-2f475173.js",
-          "routes/404-d16cca2b.js",
-          "routes/500-e47be9b7.js",
-          "routes/gists-3129f28e.js",
-          "routes/gists.mine-f34b0db7.js",
-          "routes/gists/$username-79d26e83.js",
-          "routes/gists/index-fec260b4.js",
-          "routes/index-50d745c3.js",
-          "routes/methods-a4281924.js",
-          "routes/page/four-11933b23.js",
-          "routes/page/three-fd4284d2.js",
+          "entry-server.js",
+          "pages/one.js",
+          "pages/two.js",
+          "routes/404.js",
+          "routes/500.js",
+          "routes/gists.js",
+          "routes/gists.mine.js",
+          "routes/gists/$username.js",
+          "routes/gists/index.js",
+          "routes/index.js",
+          "routes/methods.js",
+          "routes/page/four.js",
+          "routes/page/three.js",
           "server-manifest.json",
         ]
       `);
@@ -152,43 +152,43 @@ describe.skip("building", () => {
         Object {
           "entries": Object {
             "entry-server": Object {
-              "file": "entry-server-4bebe6c1.js",
+              "file": "entry-server.js",
             },
             "pages/one": Object {
-              "file": "pages/one-53b19748.js",
+              "file": "pages/one.js",
             },
             "pages/two": Object {
-              "file": "pages/two-2f475173.js",
+              "file": "pages/two.js",
             },
             "routes/404": Object {
-              "file": "routes/404-d16cca2b.js",
+              "file": "routes/404.js",
             },
             "routes/500": Object {
-              "file": "routes/500-e47be9b7.js",
+              "file": "routes/500.js",
             },
             "routes/gists": Object {
-              "file": "routes/gists-3129f28e.js",
+              "file": "routes/gists.js",
             },
             "routes/gists.mine": Object {
-              "file": "routes/gists.mine-f34b0db7.js",
+              "file": "routes/gists.mine.js",
             },
             "routes/gists/$username": Object {
-              "file": "routes/gists/$username-79d26e83.js",
+              "file": "routes/gists/$username.js",
             },
             "routes/gists/index": Object {
-              "file": "routes/gists/index-fec260b4.js",
+              "file": "routes/gists/index.js",
             },
             "routes/index": Object {
-              "file": "routes/index-50d745c3.js",
+              "file": "routes/index.js",
             },
             "routes/methods": Object {
-              "file": "routes/methods-a4281924.js",
+              "file": "routes/methods.js",
             },
             "routes/page/four": Object {
-              "file": "routes/page/four-11933b23.js",
+              "file": "routes/page/four.js",
             },
             "routes/page/three": Object {
-              "file": "routes/page/three-fd4284d2.js",
+              "file": "routes/page/three.js",
             },
           },
           "version": Any<String>,
@@ -207,19 +207,19 @@ describe.skip("building", () => {
 
       expect(getFilenames(output)).toMatchInlineSnapshot(`
         Array [
-          "_shared/Shared-f413059e.js",
-          "_shared/node_modules/@babel/runtime-f4ff0cc0.js",
-          "_shared/node_modules/@mdx-js/react-273b70e4.js",
-          "_shared/node_modules/@remix-run/react-20d518de.js",
-          "_shared/node_modules/history-45437576.js",
-          "_shared/node_modules/object-assign-c47a16a6.js",
-          "_shared/node_modules/prop-types-10e9577e.js",
-          "_shared/node_modules/react-5eb30eef.js",
-          "_shared/node_modules/react-dom-a40050cb.js",
-          "_shared/node_modules/react-is-16d950a1.js",
-          "_shared/node_modules/react-router-953744b1.js",
-          "_shared/node_modules/react-router-dom-774c78bd.js",
-          "_shared/node_modules/scheduler-0ec08ff9.js",
+          "_shared/Shared-96366458.js",
+          "_shared/__babel/runtime-88c72f87.js",
+          "_shared/__mdx-js/react-95913140.js",
+          "_shared/__remix-run/react-cde80f5e.js",
+          "_shared/history-4d09a8ec.js",
+          "_shared/object-assign-510802f4.js",
+          "_shared/prop-types-dba85e17.js",
+          "_shared/react-25ab190f.js",
+          "_shared/react-dom-366265b2.js",
+          "_shared/react-is-d37fbdde.js",
+          "_shared/react-router-d1437d00.js",
+          "_shared/react-router-dom-a453ee32.js",
+          "_shared/scheduler-2e024e35.js",
           "asset-manifest.json",
           "entry-browser.js",
           "global.css",
@@ -312,36 +312,36 @@ describe.skip("building", () => {
 
       expect(getFilenames(output)).toMatchInlineSnapshot(`
         Array [
-          "_shared/Shared-d3b72835.js",
-          "_shared/node_modules/@babel/runtime-f4ff0cc0.js",
-          "_shared/node_modules/@mdx-js/react-a6870c6c.js",
-          "_shared/node_modules/@remix-run/react-872ba3a3.js",
-          "_shared/node_modules/history-8b818913.js",
-          "_shared/node_modules/object-assign-c47a16a6.js",
-          "_shared/node_modules/prop-types-d62d43a5.js",
-          "_shared/node_modules/react-2ea82506.js",
-          "_shared/node_modules/react-dom-f1b98511.js",
-          "_shared/node_modules/react-is-08ef681b.js",
-          "_shared/node_modules/react-router-21790491.js",
-          "_shared/node_modules/react-router-dom-3dd78536.js",
-          "_shared/node_modules/scheduler-db377390.js",
+          "_shared/Shared-901c4c50.js",
+          "_shared/__babel/runtime-88c72f87.js",
+          "_shared/__mdx-js/react-7d23cc18.js",
+          "_shared/__remix-run/react-5a424f9a.js",
+          "_shared/history-e6417d88.js",
+          "_shared/object-assign-510802f4.js",
+          "_shared/prop-types-dbe6eedb.js",
+          "_shared/react-c26b1730.js",
+          "_shared/react-dom-2b0040a5.js",
+          "_shared/react-is-8eac45b9.js",
+          "_shared/react-router-10d3b386.js",
+          "_shared/react-router-dom-f0b00c95.js",
+          "_shared/scheduler-240bd53d.js",
           "asset-manifest.json",
-          "entry-browser-79f83a1b.js",
+          "entry-browser-5aae29b7.js",
           "global-ec887178.css",
-          "pages/one-5e26c998.js",
-          "pages/two-ecffc1f5.js",
-          "routes/404-05f74ff9.js",
-          "routes/500-83189ced.js",
+          "pages/one-6e4f5f3a.js",
+          "pages/two-b6fe9a86.js",
+          "routes/404-28c42fa3.js",
+          "routes/500-53a4a2d5.js",
+          "routes/gists-3ab00298.js",
           "routes/gists-6db1f83b.css",
-          "routes/gists-d4dac369.js",
-          "routes/gists.mine-65276ff1.js",
-          "routes/gists/$username-5d38ef92.js",
-          "routes/gists/index-6be545ac.js",
-          "routes/index-23df3a5b.js",
-          "routes/methods-b79e3f8c.js",
+          "routes/gists.mine-793a5f58.js",
+          "routes/gists/$username-c4ddecda.js",
+          "routes/gists/index-7ea1cb78.js",
+          "routes/index-812a1332.js",
+          "routes/methods-72ab071f.js",
           "routes/methods-e15212f5.css",
-          "routes/page/four-e83669e5.js",
-          "routes/page/three-e439ecde.js",
+          "routes/page/four-f90c3608.js",
+          "routes/page/three-97073d4a.js",
         ]
       `);
 
@@ -353,52 +353,52 @@ describe.skip("building", () => {
         Object {
           "entries": Object {
             "entry-browser": Object {
-              "file": "entry-browser-79f83a1b.js",
+              "file": "entry-browser-5aae29b7.js",
             },
             "global.css": Object {
               "file": "global-ec887178.css",
             },
             "pages/one": Object {
-              "file": "pages/one-5e26c998.js",
+              "file": "pages/one-6e4f5f3a.js",
             },
             "pages/two": Object {
-              "file": "pages/two-ecffc1f5.js",
+              "file": "pages/two-b6fe9a86.js",
             },
             "routes/404": Object {
-              "file": "routes/404-05f74ff9.js",
+              "file": "routes/404-28c42fa3.js",
             },
             "routes/500": Object {
-              "file": "routes/500-83189ced.js",
+              "file": "routes/500-53a4a2d5.js",
             },
             "routes/gists": Object {
-              "file": "routes/gists-d4dac369.js",
+              "file": "routes/gists-3ab00298.js",
             },
             "routes/gists.css": Object {
               "file": "routes/gists-6db1f83b.css",
             },
             "routes/gists.mine": Object {
-              "file": "routes/gists.mine-65276ff1.js",
+              "file": "routes/gists.mine-793a5f58.js",
             },
             "routes/gists/$username": Object {
-              "file": "routes/gists/$username-5d38ef92.js",
+              "file": "routes/gists/$username-c4ddecda.js",
             },
             "routes/gists/index": Object {
-              "file": "routes/gists/index-6be545ac.js",
+              "file": "routes/gists/index-7ea1cb78.js",
             },
             "routes/index": Object {
-              "file": "routes/index-23df3a5b.js",
+              "file": "routes/index-812a1332.js",
             },
             "routes/methods": Object {
-              "file": "routes/methods-b79e3f8c.js",
+              "file": "routes/methods-72ab071f.js",
             },
             "routes/methods.css": Object {
               "file": "routes/methods-e15212f5.css",
             },
             "routes/page/four": Object {
-              "file": "routes/page/four-e83669e5.js",
+              "file": "routes/page/four-f90c3608.js",
             },
             "routes/page/three": Object {
-              "file": "routes/page/three-e439ecde.js",
+              "file": "routes/page/three-97073d4a.js",
             },
           },
           "version": Any<String>,